Day2_MoveAllZeroesToEnd
Program to move all the zeros in the array to the right end while maintaining the relative order of the non-zero elements.

IMPORTANT TIP:

The platform expects non-static method inside a class named Solution.

The non-static method should have the return value for what you're required to return to solve the question.

You are not required to call this non-static method inside a main method. You dont need to create the main method, the code will work fine without it